+++ Vineyard/Vineyard/FirebaseDataManager.swift
@@ -7,31 +7,31 @@
 
 // import FirebaseFirestore
 
-class FirebaseDataManager: DatabaseServiceProtocol {
-    
-    func fetchGroupFromDB(groupID: String) async throws -> Group? {
-        <#code#>
-    }
-    
-    func fetchPeopleFromDB(peopleIDs: [String]) async throws -> [Person] {
-        <#code#>
-    }
-    
-    func fetchResolutionsFromDB(resolutionIDs: [String]) async throws -> [Resolution] {
-        <#code#>
-    }
-    
-    func addPersonToDB(person: Person) async throws {
-        <#code#>
-    }
-    
-    func addResolutionToDB(resolution: Resolution) async throws {
-        <#code#>
-    }
-    
-    func addGroupToDB(group: Group) async throws {
-        <#code#>
-    }
-    
-    
-}
+//class FirebaseDataManager: DatabaseServiceProtocol {
+//    
+//    func fetchGroupFromDB(groupID: String) async throws -> Group? {
+//        <#code#>
+//    }
+//    
+//    func fetchPeopleFromDB(peopleIDs: [String]) async throws -> [Person] {
+//        <#code#>
+//    }
+//    
+//    func fetchResolutionsFromDB(resolutionIDs: [String]) async throws -> [Resolution] {
+//        <#code#>
+//    }
+//    
+//    func addPersonToDB(person: Person) async throws {
+//        <#code#>
+//    }
+//    
+//    func addResolutionToDB(resolution: Resolution) async throws {
+//        <#code#>
+//    }
+//    
+//    func addGroupToDB(group: Group) async throws {
+//        <#code#>
+//    }
+//    
+//    
+//}
